chore/#81: ResponseMoodDto 변수 타입 변경

+++ app/src/main/java/com/wearerommies/roomie/data/dto/response/ResponseMoodDto.kt
@@ -30,7 +30,7 @@ data class ResponseMoodDto(
         @SerialName("isPinned")
         val isPinned: Boolean,
         @SerialName("contractTerm")
-        val contractTerm: Long,
+        val contractTerm: Int,
         @SerialName("mainImgUrl")
         val mainImgUrl: String,
     )
